+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## UNRELEASED
+FEATURES:
+
+* agent: add `local-only` parameter to operator/keyring list requests to force queries to only hit local servers. [[GH-6279](https://github.com/hashicorp/consul/pull/6279)]
+
+BUG FIXES:
+
+* snapshot: fix TCP half-close implementation for TLS connections [[GH-6216](https://github.com/hashicorp/consul/pull/6216)]
 
 BREAKING CHANGES: